requirements.txt
fastapi>=0.68.0,<1.0.0
uvicorn>=0.34.0,<1.0.0
sse-starlette>=2.2.1,<3.0.0
watchdog>=6.0.0,<7.0.0
python-dotenv>=1.0.1,<2.0.0
google-generativeai>=0.3.0,<1.0.0
httpx>=0.28.1,<1.0.0
mcpadapt>=0.0.15,<1.0.0 